DataModules: run all data augmentation on the GPU (#992)
* DataModules: run all data augmentation on the GPU

* Passing tests

* Update BigEarthNet

* Break ChesapeakeCVPR

* Update COWC

* Update Cyclone

* Update ETCI2021

* mypy fixes

* Update FAIR1M

* Update Inria

* Update LandCoverAI

* Update LoveDA

* Update NAIP

* Update NASA

* Update OSCD

* Update RESISC45

* Update SEN12MS

* Update So2Sat

* Update SpaceNet

* Update UCMerced

* Update USAVars

* Update xview

* Remove seed

* mypy fixes

* OSCD hacks

* Add NonGeoDataModule base class

* Fixes

* Add base class to docs

* mypy fixes

* Fix several tests

* Fix Normalize

* Syntax error

* Fix bigearthnet

* Fix dtype

* Consistent kornia import

* Get regression datasets working

* Fix detection tests

* Fix some chesapeake bugs

* Fix several segmentation issues

* isort fixes

* Undo breaking change

* Remove more code duplication, standardize docstrings

* mypy fixes

* Add default augmentation

* Augmentations can be any callable

* Fix datasets tests

* Fix datamodule tests

* Fix more datamodules

* Typo fix

* Set up val_dataset even when fit

* Fix classification tests

* Fix ETCI2021

* Fix SEN12MS

* Add GeoDataModule base class

* Fix several chesapeake bugs

* Fix dtype and shape

* Fix crs/bbox issue

* Fix test dtype

* Fix unequal size stacking error

* flake8 fix

* Better checks on sampler

* Fix bug introduced in NAIP dm

* Fix chesapeake dimensions

* Add one to mask

* Fix missing imports

* Fix batch size

* Simplify augmentations

* Don't run test or predict without datasets

* Fix tests

* Allow shared dataset

* One more try

* Fix typo

* Fix another typo

* Fix Chesapeake dimensions

* Apply augmentations during sanity check too

* Don't reuse fixtures

* Increase coverage

* Fix ETCI tests

* Test predict_step

* Test all loss methods

* Simplify validation plotting

* Document new classes

* Fix plotting

* Plotting should be robust in case dataset does not contain RGB bands

* Fix flake8

* 100% coverage of trainers

* Add lightning-lite dependency

* Revert "Add lightning-lite dependency"

This reverts commit 1df7291ae59f6257a2cabd20a6c767e178bf4f0f.

* Define our own MisconfigurationException

* Properly test new data module base classes

* Fix mistake in setup call

* ExtractTensorPatches runs into OOM errors

* Test both fast_dev_run True and False

* Fix plot methods

* Fix OSCD tests

* Fix bug with inconsistent train/val/test splits between stages

* Fix issues with images of different sizes

* Fix OSCD tests

* Fix OSCD tests

* Bad rebase

* No trainer for OSCD so no need for config

* Bad rebase

* plot: only works during validation

* Fix collation of NASA Marine Debris dataset

* flake8 fix

* Quick test

* Revert "Quick test"

This reverts commit f465efcbef904b8a5bc2257f2800eed931c491ab.

* 56 workers is a bit excessive

+++ tests/datasets/test_geo.py
@@ -211,7 +211,7 @@ class TestRasterDataset:
         x = custom_dtype_ds[custom_dtype_ds.bounds]
         assert isinstance(x, dict)
         assert isinstance(x["image"], torch.Tensor)
-        assert x["image"].dtype == torch.int64
+        assert x["image"].dtype == torch.float32
 
     def test_invalid_query(self, sentinel: Sentinel2) -> None:
         query = BoundingBox(0, 0, 0, 0, 0, 0)
